Schedule Cron Week
Create Meteor Collection for Schedules in folder lib/schedule and allow crud operations.
Update form view to be able to create the schedules.
Take the checkbox status for days of the week and hours of the days save them into arrays.
Create Method to insert into Schedule Collection.
Create Publish to retreive data from the Schedule Collections.
Review and update the ER in the diagram, schedule attributes.
Next: Testing the schedule insert form.