Schedule Cron Weekly Form and List table of Schedules.
I change the schedule Schema the hours array is not of booleans, now is an array of Strings, then I save the Schedule form into the collection and show the list of Schedules in a table, In the schedule table show the name of the actions that will perform every schedule.
Next: FIx error when reset the Form after save in db.
Next: Study RabbitMQ and activate every actions in each schedule acording to the day and hour.