Domains and Black/Lists Modules.
Created Front-end for save new domains form and table to show the list of domains created, created the publish and methods on server side.
Created the front-end and backend for to create the black and white lists of domains.
Next: Add documenation JSDoc, fix schedule and add the functionality to save the schedules.