Finish Schedule Module
Load the days, hours, the black and white list of domains created for the user.
Load all the actions available for all the users.
Then save the name, description, days, hours, The user select the actions, white and black list to be used for the actions in the schedule.
Next step: Take the schedule and activate actions acording to days and hours, and see if the actions need a blacklist or whitelist to be passed with the accounts.