Delete operation for modules Actions-Domains-Lists-Schedules.
Add delete operation but need to improve validations, when A domain is in a list and a schedule is using the list, the domain can't be deleted, or the lists.
The package used to perform update operations did not work, so we need to implement that operations with another package.
Next: Merge with fernando branch and add update operations.