Testing all actions Bots through TobMaps App.
Create a table that allow us to activate each bot just for testing purposes, to study if the bot works the same as running normal, also we fetch the emails to be used and pass the domains as args to the bot.
At the end we save all the stdout of the bot in a collection called loggers.
Now: I am testing the bots used to move from spam emails which domain fetch with domains passed as args to the bot, the bots from Outllok and Yahoo.
The Update ooperation keep to be done in another way, because for some specific modules doesn't work the implemented package.
Next: Keep testing the bots work normal using TobMaps.