Update Module Actions and Test Bots
Change the form of actions to improve save.
Fix the delete that has an issue when click on button.
Test outlook bots with TobMaps App meteor.
I found a problem with the bot that send many emails. Need to fix it to work with TobMaps as normal.
Next: Fix 4 Bot, and start to work with the Schedule to activate the actions.