Review Outlook Bots, and Deployment Steps for production and Development
I checked that the scripts of the bots work and run from MeteorJS, I fix some error that came when I was testing bots, now they are all working and can be activated with a button.
I create a file in SetupServer called DeploymentSteps.txt where are the steps to follow to deploy to production and development.
Now: I am going to create function that are going to have the actions that the user can perform, so the user can define how want the actions to be performed, the order, we can now the order and call the function the user wants to perform.