Definable Actions in Module Actions
Now the user can define the order by selecting the actions to be performed, once the user save, a meteor autoform hooks listen the submit event and call the meteot method to insert the new actions.
I documented with JSDoc the module Actions.
Next: Organize the CasperJS Bots in a single script using function.