Create nodejs modules of the CasperJS Script Actions
I added hooks.js file to domain,list, action and schedule module to be available to call meteor method and use toastr, I also clean the form when click on reset button. On the casperjs script I am Using exports.module and require, we clean each casperjs script and modularize everything so we can import in the main script all the modules with require.
Next: Test the casperjs modules in the main casperjs script, passing actions to be activated and all the data from the schedules.