Schedule Module - Table View and Activities View.
The table that shows each schedule and its actions has an error in a meteor helper function that loads the names of the actions that will perform the schedule, this is because we made some changes to be available to define our actions, the definable actions, so I change the template and helper to be available to show every actions that the schedule will perform in the order the user has decided.
Then once this got fixed I modify the files routes,js, navigation.html and navigation.js to use menu to show options.
Next work: I am going to checkout to branch bots module to keep working on the bots, creating functions and use that funtions.