CasperJS Modules
After Testing the casperjs bots as node modules and didn't work, we create and work with functions to develop modules to be more easy to work and mantain.
When testing with 4 actions, some errors start to appear when I was clicking to logout and login in the next account, I solve this error calling the direct link to logout.
Next: I join yahoo functions into the main script and test both Yahoo and Outlook.